refactor some pydantic to their new places from api
