+++ nginx/nginx.conf
@@ -4,8 +4,6 @@ server {
 
     resolver 127.0.0.11 valid=30s;
 
-    # --- Rule 0: Serve the static consolidated OpenAPI file ---
-
     location = /form-manual {
         alias /usr/share/nginx/html/form_manual.html;
         default_type text/html;
