+++ services/api_service/src/app/schemas/forms.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class FormSchemaUpload(BaseModel):
+    version: str
+    schema_data: dict
