+++ services/api_service/src/app/schemas/sessions.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+
+class TelegramSessionRequest(BaseModel):
+    telegram_id: int
+
+
+class SessionResponse(BaseModel):
+    application_uuid: str
